refactor(core): avoid circular dependency in ripples
Moves some types around in order to avoid a circular dependency between the ripple renderer and ripple ref.
