+++ src/material/core/ripple/ripple-ref.ts
@@ -6,11 +6,29 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {RippleConfig, RippleRenderer} from './ripple-renderer';
-
 /** Possible states for a ripple element. */
 export const enum RippleState {
   FADING_IN, VISIBLE, FADING_OUT, HIDDEN
+}
+
+export type RippleConfig = {
+  color?: string;
+  centered?: boolean;
+  radius?: number;
+  persistent?: boolean;
+  animation?: RippleAnimationConfig;
+  terminateOnPointerUp?: boolean;
+};
+
+/**
+ * Interface that describes the configuration for the animation of a ripple.
+ * There are two animation phases with different durations for the ripples.
+ */
+export interface RippleAnimationConfig {
+  /** Duration in milliseconds for the enter animation (expansion from point of contact). */
+  enterDuration?: number;
+  /** Duration in milliseconds for the exit animation (fade-out). */
+  exitDuration?: number;
 }
 
 /**
@@ -22,7 +40,7 @@ export class RippleRef {
   state: RippleState = RippleState.HIDDEN;
 
   constructor(
-    private _renderer: RippleRenderer,
+    private _renderer: {fadeOutRipple(ref: RippleRef): void},
     /** Reference to the ripple HTML element. */
     public element: HTMLElement,
     /** Ripple configuration used for the ripple. */
