+++ tools/public_api_guard/material/core.d.ts
@@ -366,7 +366,9 @@ export declare class RippleRef {
     config: RippleConfig;
     element: HTMLElement;
     state: RippleState;
-    constructor(_renderer: RippleRenderer,
+    constructor(_renderer: {
+        fadeOutRipple(ref: RippleRef): void;
+    },
     element: HTMLElement,
     config: RippleConfig);
     fadeOut(): void;
